[happy @ 2000-12-03 16:21:50 by simonmar] Happy can now provide multiple entry points to a single grammar. The syntax is like this:
 	%name parseModule module
	%name parseExpr   expr

	%%
	module : ...
	expr   : ...

There are some type changes; before you used to be able to give the
lexer the type ((Token -> P T) -> P T) where T is the result type of
the parser.  Now it should be polymorphic, i.e. ((Token -> P a) -> P
a), because the parser can have several result types.  (Aside: I'll
probably change the type of the lexer to simply (P Token) before the
next release; the continuation style is of dubious merit).  The type
of happyError must now also be (P a) rather than (P T).

This change involved much fiddling around generalising Happy's deeply
wired in notions of what "token zero" and "state zero" mean, what the
value of the first non-terminal is, etc.  All the tests still work,
though.

Documentation updates to follow.

I also did a cleanup sweep while I was in the code, and took advantage
of -fwarn-unused-imports and -fwarn-unused-binds.