[happy @ 2001-09-24 15:48:54 by simonmar] Previously, tokens which were not declared to have a precedence/associativity were given a default precedence lower than everything else and made non-associative.
This doesn't match what bison does and seems confusing (at least it
confused me).  So I'm changing the behaviour so that you have to
declare all precedences explicitly, and a conflict between an operator
with a declared precedence and one without will be reported as a
conflict rather than being silently resolved.