[happy @ 2003-02-12 16:36:23 by simonmar] Another fixed-point.
I've implemented GHC-specific hacks to reduce the size of the tables,
using exactly the same technique as in Happy, i.e. the tables are
encoded as strings.  Also some unboxery was used in the lexer engine
itself for speed.  The GHC-specific stuff is selected with the -g
flag, as for Happy.

This reduces the size of the binary of Alex's own lexer by a further
80% or so.