[happy @ 2003-02-18 17:13:46 by simonmar] Bootstrap fixed-point.
We now provide a monad-based scanner engine, which is more flexible in
some ways (and less flexible in others) than the previous
continuation-based method.  The monad version is selected with the -m
command-line flag.

Alex's own lexer now uses the monad engine, and it now does a better
job of interpreting code fragments which contain braces, including
within strings and character literals.