[happy @ 2003-03-31 16:52:21 by simonmar] Some changes to the programmer's interface to a generated scanner.
Now you get a bare-bones scanner where Alex supplies

	type StartCode
	alexScan :: AlexInput		 -- The current input
		 -> StartCode		 -- The "start code"
		 -> Maybe (           	 -- Nothing on error or EOF
			     AlexInput,  -- The remaining input
			     Int,        -- Length of this token
			     action      -- The action (unknown type)
			  )

You can use this scanner by supplying the following:

	type AlexInput
	alexGetChar  :: AlexInput -> Maybe (Char,AlexInput)
	alexPrevChar :: AlexInput -> Char

or there is a selection of "wrappers" which provide suitable
definitions for the above.  eg. saying

  %wrapper "monad"

in the grammer file gets you a default monad which keeps track of the
input and can handle parse errors, and

  %wrapper "gscan"

gets you an interface similar to the old Alex.

Or alternatively you can build your own monad on top of the bare-bones
scanner, which is what Alex itself now does.  As a result it can now
report lexical errors in the grammar file properly.