fix the GLR parser generator feature of happy
When generating GLR parsers, happy attempted to mix user code from the .y file,
delitted .lhs source, and happy generated code together, this was buggy causing
the resulting programs to not compile and happy to not be able to include the
GLR_Lib properly. This patch fixes this by only mixing non-literate haskell
code into the happy generated GLR parser. 

This patch also fixes bugs in a few of the glr examples and replaces uses of
FiniteMap with Data.Map as needed.