On parse error, allow to report a list of possible next tokens
This change introduces the %errorsig attribute, that changes
the API of the provided handler function. A [String] of
possible next token can be passed to the function in order
to provide a better error message to the user.

We need to pass over the list of possible token names to the
happyFail function. For example:

    action_12 (12) = happyShift action_9
    action_12 (13) = happyShift action_10
    action_12 (14) = happyShift action_11
    action_12 (15) = happyShift action_12
    action_12 (5) = happyGoto action_15
    action_12 (6) = happyGoto action_5
    action_12 (7) = happyGoto action_6
    action_12 (8) = happyGoto action_7
    action_12 (9) = happyGoto action_8
    action_12 _ = happyFail ["idT","numT","boolT","\"(\""]

Signed-off-by: Dan Aloni <dan@kernelim.com>