Wrap only numbers for offset tables
Added new variable 'happyMinOffset' that is present only when arrays are enabled.
This variable keeps track of the lowest offset value we should ever find in the
values of 'happyActOffsets' and 'happyGotoOffsets'. That way, if we ever find a
value that is smaller, we know it accidentally wrapped around (and we can correct
for the wraparound).

Since the offset tables contain only small negative numbers (as many as you have
rules), this effectively doubles the size of offsets supported.