Support larger for offset tables
Added new variable 'happyMinOffset' that is present only when arrays are enabled.
This variable keeps track of the lowest offset value we should ever find in the
values of 'happyActOffsets' and 'happyGotoOffsets'. That way, if we ever find a
value that is smaller, we know it accidentally wrapped around (and we can correct
for the wraparound).

Added a check so that if Happy is run with '--ghc' and the offset tables are
actually too large, Happy will fail with an error message instead of generating
bogus code.

Since the negative numbers in offset tables are generally quite small, this
effectively doubles the size of offsets supported.