[ fixed #161 ] .info: don't drop all parse items when printing a state
Items that start with a dot are usually added by completion
and thus redundant and dropped from the printout.
However, if the initial item started with a dot, it should not be dropped,
otherwise there will be no items left.  Thus, should there be no items
not starting with a dot, we print the initial item.