No bootstrapping
Before this patch, building 'happy' required a pre-built binary of
'happy'. This was elegant in the same way a self-hosting compiler is
elegant. But it also made building purely from source more complicated
than needed.

This patch introduces a small, bespoke parsing library, and applies it
for parsing .y and .ly files.

Now 'happy' doesn't depend on itself, and can be built using just GHC.