Optional bootstrapping (#175)
Before this patch, building 'happy' required a pre-built binary of 'happy'. This was elegant in the same way a self-hosting compiler is elegant. But it also made building purely from source more complicated than needed.

Now, we have a subset of the functionality also written with parser combinators. One can build that minimal happy from source, and then bootstrap the regular one from it. This simplifies the build process considerably, along with helping with various (albeit mostly theoretical) trust issues.

The parser combinators are bespoke now, but since they need not be performant we will probably just switch to `ReadP` from base, later.

Co-authored-by: Vladislav Zavialov <vlad.z.4096@gmail.com>