Split `BookendedAbsSyn` out from `AbsSyn`
`BookendedAbsSyn` wraps it to separate the information needed by the
middle parts of happy from the header and footer which are just added
back at the end.

The code looks nicer to me now, with fewer wildcard and @-patterns,
therefore I conclude that this is makes the data structures more natural
for what we are doing.