Split out happy-codegen-common
There was some extra information stuff in `Grammar` which had nothing to
do with the grammar, but was simply there because `Grammar` was also
playing the role of capturing all information from the abstract syntax.

I don't think that's good. As we really try to really make libraries out
of this stuff we should be stricter and stricter about separating
concerns. Grammar should really just be that, the grammar, and the code
in `happy-tabular` should not be privy to information that is just for
the backend.

Splitting out a code generation `CommonOptions` type is a first step to
rectifying this. I hope we can do a few more refactors like this to
really make each data type shine on its own.

---

I don't want to sound to harsh though, since we purposely made the split
low impact with more cleanups -- such as this -- left for later. It is
of course easier to see what's good and what's bad once the code is
split up!

Also, this change calls into question my previous `BookendedAbsSyn`.
We're now acknowledging the abstract syntax mixes "middle" and back end
concerns, and those are not properly separated until the nest step.
Given that, there isn't much use of making `BookendedAbsSyn` when we
could just stick the header and footer in `CommonOptions`.