CI to build happy via pre-installed happy
After installing the prerequisites which include a released version of happy,
this CI runs:

    make sdist
    make sdist-test
    cabal build

`cabal build` needs to run after `make sdist`
as the latter creates the contents of the data/ directory.

This CI took elements of Haskell-CI and Agda's CI.
It uses haskell/action/setup to install ghc and cabal.

Restrict this CI to pushes and PR to branches
- master
- anything starting with "ci" (for running CI on branches without PR)
- anything starting with "release"