Begin to separate grammars from their eliminators
Instead of assuming we always have raw Haskell code (a `String`) which
eliminates each production rule, use a type variable.

As one would hope, much code only cares about the grammar itself, and
works with that type abstract. The new types thus recover free theorems,
yay! Only a little bit of the frontend and the backends actually wants
it to be strings.

These type parameter can be leveraged in the future. Some possibilities
are:

- a TH-based frontend to Happy

- interpreting the grammar where we don't have custom elimination rules
  at all, but just create an "untyped" parse tree.