Move Paths_happy_lib into `tabular` to prevent a Cabal bug concerning --libsubdir (#328)
Not only do we need to move the module, we also need to reexport it
under a different name lest we get an unhelpful error suggesting to
add `Paths_happy_lib` to the other-modules section of
`backend-{lalr,glr}`.

It is likely that this change will also fix `cabal v1-install happy` (#315).

Fixes #328.